Use --symlink-type=sys for Cygwin setup
Cygwin setup 2.912 has a bug in native Windows symlink support and
writes some corrupt symlinks; in particular /etc/ssl/certs is corrupt
which breaks git and curl.

Depexts will still install using native Windows symlinks, but this
should hopefully be less of an issue as the bug affects absolute
symlinks, which are less common outside of the base packages.