Define CLICOLOR_FORCE=1 env var in CI runs
This env var is supported by some tools to force them to emit ANSI
escape sequences corresponding to colors even if they detect that
they're not running in tty, which is th case for CI jobs.

It is going to be supported in Dune 3.6.0, and perhaps other compilers
and build systems.

Signed-off-by: Antonin Décimo <antonin@tarides.com>